move user and userdb to new location
