+++ src/eduid/userdb/maccapi/__init__.py
@@ -1,0 +1,1 @@
+from eduid.userdb.maccapi.userdb import ManagedAccount, ManagedAccountDB
